+++ src/main/java/com/smu/householdaccount/repository/BoardLikeRepository.java
@@ -7,9 +7,7 @@ import java.util.Optional;
 
 public interface BoardLikeRepository extends JpaRepository<BoardLike, Long> {
 
-    // 회원이 특정 게시글에 좋아요 했는지 확인
-    Optional<BoardLike> findByPost_IdAndMember_MemberId(int postId, String memberId);
+    Optional<BoardLike> findByPost_IdAndMember_MemberId(Long postId, String memberId);
 
-    // 특정 게시글의 좋아요 개수 조회
-    Long countByPost_Id(int postId);
+    long countByPost_Id(Long postId);
 }
